# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sensor Predictive Backend API
  version: 0.1.0
  description: Backend MVP for incident prediction in engineering collectors.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      security: []
      tags: [Authentication]
      summary: Authenticate through the configured identity provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password, minLength: 12 }
      responses:
        "200":
          description: JWT access token and current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401": { description: Invalid username or password }
  /api/v1/auth/me:
    get:
      tags: [Authentication]
      summary: Get identity encoded in the current access token
      responses:
        "200":
          description: Current identity
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: { type: string, format: uuid }
                  username: { type: string }
                  role: { $ref: "#/components/schemas/Role" }
        "401": { description: Missing, invalid, or expired token }
  /api/v1/users:
    get:
      tags: [Administration]
      summary: List local users (admin only)
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: User page
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/User" }
        "403": { description: Admin role is required }
    post:
      tags: [Administration]
      summary: Create a local user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, role]
              properties:
                username: { type: string }
                password: { type: string, format: password, minLength: 12 }
                role: { $ref: "#/components/schemas/Role" }
      responses:
        "201":
          description: Created user without password data
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { description: Invalid role or password }
        "409": { description: Username already exists }
  /api/v1/audit-logs:
    get:
      tags: [Administration]
      summary: List security and mutation audit entries (admin only)
      parameters:
        - in: query
          name: username
          schema: { type: string }
        - in: query
          name: action
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Audit page
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AuditEntry" }
        "403": { description: Admin role is required }
...
  /api/v1/imports:
    get:
      tags: [Imports]
      summary: List CSV import jobs
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Import job page
  /api/v1/imports/{import_id}:
    get:
      tags: [Imports]
      summary: Get CSV import progress
      parameters:
        - in: path
          name: import_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Import job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportJob" }
        "404": { description: Import job not found }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      in: query
      name: limit
      schema: { type: integer, default: 100, maximum: 500 }
    Offset:
      in: query
      name: offset
      schema: { type: integer, default: 0, minimum: 0 }
...
  schemas:
    Role:
      type: string
      enum: [admin, dispatcher, analyst, manager]
    User:
      type: object
      required: [user_id, username, role, active, created_at]
      properties:
        user_id: { type: string, format: uuid }
        username: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        active: { type: boolean }
        created_at: { type: string, format: date-time }
        last_login_at: { type: string, format: date-time, nullable: true }
    LoginResponse:
      type: object
      required: [access_token, token_type, expires_at, user]
      properties:
        access_token: { type: string }
        token_type: { type: string, enum: [Bearer] }
        expires_at: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/User" }
    AuditEntry:
      type: object
      required: [audit_id, username, action, resource_type, method, path, details, created_at]
      properties:
        audit_id: { type: integer, format: int64 }
        user_id: { type: string, format: uuid }
        username: { type: string }
        action: { type: string }
        resource_type: { type: string }
        resource_id: { type: string }
        method: { type: string }
        path: { type: string }
        remote_ip: { type: string }
        details: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
...
    ImportJob:
      type: object
      properties:
        import_id: { type: string, format: uuid }
        import_kind: { type: string, enum: [objects, channels, events] }
        file_name: { type: string }
        status: { type: string, enum: [pending, running, completed, failed] }
        processed_rows: { type: integer, format: int64 }
        failed_rows: { type: integer, format: int64 }
        error_message: { type: string }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }